Fixes bug with compressed backups
When the compress backups option was set, the directory that the backup
was based on was not removed after it was compressed.  This lead to the
backups directory being poluted with duplicate backups, and the the sort
that happened grouped the duplicates together, so only half of the
requested backups were actually kept.